Change the terms "operation" and "sender operation" back to "algorithm" and "sender algorithm", as "operation" is already used to refer to the work performed by a sender, and the "operation" terminology conflicts with the notion of operation states.
Keep the "sender consumer" term and roll it out everywhere.

Prefer being more explicit and saying "sender algorithm" in a bunch of places
to avoid mental clashes with the (parallel) sequence algorithms (e.g. `<algorithm>`).

Make sure we only use "operation" and "algorithm" in the places where we really
mean it.

Fixes #29. Fixes #85.