Simplify examples
Currently the examples are really hard to read. We have many lambdas that capture by reference and do complicated things. However, the actual intent for the examples is to show how to build an asynchronous pipeline.

Therefore, pull the lambdas out, give them (hopefully) reasonable names and lets the reader focus on what is important