Wording defect on the "transfer" algorithms w.r.t. get_completion_scheduler
The working implies that get_completion_scheduler for all 3 CPOs will
return a scheduler equivalent to the one received as an argument. But
this can be only guaranteed in the case of set_value_t and set_done_t,
and cannot be guaranteed in case of set_error.

If an exception is thrown while trying to use the given scheduler to
schedule work, then the specification indicates that `set_error` is
called directly, i.e., so it may not be on the given scheduler.

This fix applies to `transfer`, `schedule_when` and `transfer_when_all`