Fix error_types and sends_stopped for the transfer `family` of algos
The `schedule_from` algo has its own completion signals computation that
joins the `error_types` and `sends_stopped` from the input scheduler and
the sender created by the given scheduler.

This fixes also the bad behavior of `transfer` and `transfer_just`, and
helps on fixing some tests based on these.

Adding more tests for completion signatures and `__transform` in the
process of chasing a hard-to-find bug that manifested on downstream
senders (using the `Variant` template parameter multiple times in the
`error_types` definition; depending on the transformation passed to
`error_types`, that leads to strange errors).