Fix customization of `let_error`
When checking if there is a customization (with scheduler) for
`let_error` the paper specifies that `get_completion_scheduler<set_value_t>`
should be called. Previously, the `set_error_t` tag was used.

This applies also to `let_stopped`, but in that case the test was
working by coincidence (the completion sender was the same for
`set_value_t` and for `set_stopped_t`).