Fix occasional deadlock in sync_wait test.
The problem was in the `impulse_scheduler` code and its assumptions.
Previously, `impulse_scheduler` assumed that the sender it creates will
be connected to a receiver before `start_next()` was called. Otherwise,
`start_next()` would do nothing, and the operation will never be started.

This adds basic synchronization in `impulse_scheduler` to allow these
two operations to be done independently. Now, if `start_next()` is
called when no sender was connected, this will wait for the sender to be
created. This also protects the access to the vector used to store the
commands to be executed by sender's operation_state.

Fix test that calls an extra `start_next()`, to avoid it blocking forever.