Work in progress: aligning the graph scheduler to the specification.
This is a not-yet-working patch. It entirely breaks the distributed
scheduler, and does not yet fully compile in the maxwell example because
ptxas ends up referencing std::exception_ptr symbols; this will be fixed
in the future. It also doesn't really tackle either repeat_n or
halo_exchange yet.

This also currently uses a suboptimal strategy for sync_wait; to make it
more optimal, I will pull out the important bits of pipeline_end out of
that sender into a proper abstraction that can also be used by a
sync_wait specialization which ends up calling cudaEventSynchronize
after a graph launch, instead of launching a graph that includes a host
graph node which jumps to a trampoline execution context (which is
necessary for the non-sync_wait cases).

Missing are also specializations for the let_* family (not entirely sure
how and if to specialize them at this point), and a specialization of
ensure_started (that also requires additional exploration to figure the
correct form of it out).

schedule() is now implemented in terms of just() | schedule_from(), to
reduce the amount of repeated code. We can change this in the future if
needed.

Finally, the implementations of bulk and of then have been folded into a
single type, because most of the code ends up being identical across the
two, with changes that can be easily abstracted over.