Add monostate to schedule_from's storage
In the case where a non-default constructible type is sent in the
set_value channel and set_stopped is not sent by the predecessor to
schedule_from the first variant in the storage used by schedule_from
would not be default constructible and the whole variant would not be
default constructible. Adding an explicit monostate makes it default
constructible.