task: fix leak upon stops (#613)
When a task is awaiting a Sender which stops, the task_awaitable is
never await_resume'd (because that would resume the task).  In that
case the task_awaitable is destructed and the coroutine_handle is
leaked.  Fix this by destroying up any residual coroutine_handle in
the task_awaitable destructor.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>