make the io_uring test compile faster with nvc++ (#873)
With this change, the time to compile test_io_uring.cpp with nvc++ drops
from 20 minutes (!) to 20 seconds.

The test was clearly hitting some pathological behavior in nvc++. That
still warrents and investigation. I suspect it's due in part to the lack
of memoization for alias template specializations, causing the compiler
to re-evaluate the same compile-time computation time and again,
possibly exponentially.