Implement the iterate sequence sender (#990)
This algorithm is a sequence sender factory. It takes an input range as function parameter and returns a lock step sequence sender that iterates through the range. The sequence operation completes if the end of the range is reached or if a next sender completes with via the stop signal.

---------

Co-authored-by: Eric Niebler <eniebler@nvidia.com>