Disable io_uring tests on Linux kernel < 5.4
Some Linux distributions have <linux/io_uring.h>, but fail with
"Operation not permitted" on
	syscall(__NR_io_uring_setup, __entries, &__params)

To work around this, skip the entire test when running on a system with
a Linux kernel older than 5.4.  (I am not certain that this is the
correct check.  But it is the best I could come up with in testing on
our systems.)