Fix race condition when destroying implementation operating state.
On the default implementation side, after sending the completion message
we called the destructor of the inner operation state. This destructor
was called on the thread that completed the work (most often, not the
main thread). But, the main thread, after receiving the completion
signal may delete the operation state on the "host" side, effectively
destroying the object on the implementation side.

To solve this issue, we need to make the "host" side tell the
implementation side when it needs to delete its internal state objects.
Added new API for this. With this change, the implementation operating
state is deleted on the main thread.