Add Standalone implementation for Asio.
The idea is that let's have an asio thread pool rather then a boost.
In this way the standalone asio can be also used as an implementation.

Which implementation is used for Asio is controlled by the cmake
parameter STDEXEC_ASIO_IMPLEMENTATION. It can be 'boost' or
'standalone'.

Cmake generates a configuration file that creates a namespace:
execpools::asio_impl. This points to the corresponding implementation.
In this way the pool implementation is referenced only to this
namespace.

The configuration file is generated into the build directory, inorder to
not pollute the source tree with generated files.