Fix race in async_scope
In the original logic, when the future receiver completion is called,
`__save_completion` and `__dispatch_result` both acquire and release
the mutex. If the sequence is

 * Thread 1 calls `__save_completion`, sets `__data_`, then returns
   successfully.
 * Thread 2 calls the `__future` destructor, which acquires the mutex
   and sees that `__data_` is set, so it returns early, and since it
   still owns `__state_`, the future state is destroyed.
 * Thread 1 calls `__dispatch_result`, and attempts to access now
   freed memory.

It's not clear why `__save_completion`/`__dispatch_result` acquire and
release the mutex separately, so I combined the calls to both be done
under the mutex.

Additionally, there appeared to be a bug when setting the value/error into
`__data_` raised an exception where `__dispatch_result` was never invoked,
leading to the program to never complete. I added a test that waits
indefinitely without this fix.

Fixes #1413