Fix build issues during import.
When stdexec is used as a 3rd-party dependency for example with fetch
content it is downloaded into the cmake build folder under the _deps.

A setup like that will results a build directory structure like
build/_deps/stdexec-src
           /stdexec-build

in this context the CMAKE_BINARY_DIR is the build folder and the
CMAKE_CURRENT_BINARY_DIR is the build/_deps/stdexec-build.

Issue #1:
Right now the build will place two files into the 'main' build folder:
 - execution.bs
 - RAPIDS.cmake

It might causes problems when the main cmake project uses RAPIDS
with a different version.

Issue #2: There is no such a file stdexec_version_config.hpp under the
main build tree. And it will cause a configuration error.