fix UB from pointer-interconvertibility issue in `__detail::__op_state` implementation (#1478)
The current implementation casts from a pointer to the first member of a struct (`__state_box::__state_`) to a pointer to the containing struct (`__state_box`), but that is only valid if `__state_box::__state_` is standard layout, which is not guaranteed.

This PR changes `__state_box::__state_` into a byte array, into which the state object is emplaced. That makes `__state_box` a standard layout type unconditionally, which makes the cast from the member to the enclosing struct well-defined.