System scheduler R6 work (#1464)
* Minor fixes

* Use generic functions for replaceability.

One can replace different specialization of `query_system_context`.
The runtime setter for system context (`set_system_context_backend`) is visible and can be specialized.

* Add support for runtime environment to be passed to the backend.

* Allow inplace_stop_token to be passed to system context backend.

* Implement replaceability with a Phoenix singleton.

* Use libdispatch backend if available.

* Improve handling of type-erased receiver environment.

Instead of having a dedicated parameter of type `env`, the receiver
environment is now exposed thorough `receiver`.
We don't need to store any extra data in the operation state, and the
receiver can be queried at any time for receiver environment properties.

* Fix include problem.

* Use a spin lock instead of a mutex in `__instance_data`

Hoping that msvc will allow it to be `constinit`.

---------

Co-authored-by: Eric Niebler <eniebler@nvidia.com>