Parallel scheduler -- incorporate feedback from LEWG in Hagenberg (#1485)
* Minor fixes

* Use generic functions for replaceability.

One can replace different specialization of `query_system_context`.
The runtime setter for system context (`set_system_context_backend`) is visible and can be specialized.

* Add support for runtime environment to be passed to the backend.

* Allow inplace_stop_token to be passed to system context backend.

* Implement replaceability with a Phoenix singleton.

* Use libdispatch backend if available.

* Improve handling of type-erased receiver environment.

Instead of having a dedicated parameter of type `env`, the receiver
environment is now exposed thorough `receiver`.
We don't need to store any extra data in the operation state, and the
receiver can be queried at any time for receiver environment properties.

* Fix include problem.

* Rename the backend `system_scheduler` into `parallel_scheduler_backend`.

* Replace `query_system_context` template with `query_parallel_scheduler_backend` plain function.
Same for the setter, but also mark it as "out of spec".

* Rename `get_system_scheduler` -> `get_parallel_scheduler`.
(and associated types)
Anonymized some types that are not named by the design.

* Various small fixes to align with LEWG feedback.

* Use a spin lock instead of a mutex in `__instance_data`

Hoping that msvc will allow it to be `constinit`.

* Apply clang-tidy. Add `get_system_scheduler` as deprecated.

---------

Co-authored-by: Eric Niebler <eniebler@nvidia.com>