exec::storage_for_completion_signatures
Certain algorithms, for example when_all, must store the completions of
child operations and then later examine or forward them. The typical way
of doing this is via decay-copy into a tuple, but this erases the
reference-ness of values and errors.

The class template added by this commit, exec::
storage_for_completion_signatures, makes it convenient to store,
examine, and forward the completions of some asynchronous operation,
while also being reference-aware.