asioexec::completion_token: Initiating Function Throw Edge Case
The start member function of the operation state synthesized when using
asioexec::completion_token previously assumed that an exception thrown
from an initiating function meant that initiating function had no effect
(i.e. that there was no asynchronous operation outstanding). This is not
necessarily the case and if it's not the case it's important to
initialize the stop callback so cancellation requests can be transmitted
(as evidenced by the newly-added unit test).