asioexec::completion_token:: & ::use_sender::as_default_on
When an Asio asynchronous operation is parameterized by an executor or
an I/O object it can adopt the default completion token of that executor
or of the executor associated with that I/O object, if any. Adopting a
default completion token means that, if the initiating function does not
receive an explicit completion token parameter, it will be implicitly
parameterized by a default constructed instance of that default
completion token type.

Given that Asio I/O objects feature a rebind_executor nested template it
follows from this that one can rebind an I/O object to be associated
with an executor which exposes a certain default completion token. See
for example boost::asio::use_awaitable_t and asio::awaitable_t (in
Boost.Asio and standalone Asio, respectively). Therefore upgraded
asioexec::completion_token and ::use_sender to support the pattern
embodied by the aforementioned types from Boost.Asio/standalone Asio.
By using the newly-added as_default_on members thereof an I/O object may
be rebound to an executor type which causes that I/O object's initiating
functions (and free initiating functions parameterized thereby) to yield
a sender when no completion token is explicitly provided thereto.

See the newly-added unit tests for examples.