write_env: No Evaluated Context for Environment Type
The get_env implementation for stdexec::write_env is supplied via a
lambda. Prior to this commit that lambda relied on automatic return type
deduction. When a function relies on automatic return type deduction the
compiler must evaluate the body thereof to determine the type returned
thereby. A consequence of the aforegoing formulation is that if any
statement in the body of the lambda directly or transitively relied on
the completeness of an incomplete type compilation would fail.

The above might seem normal and non-problematic until one considers that
it is fairly normal in C++ code to compute (and rely upon) types and
other properties of functions long before any bona fide evaluation of
that function occurs. This motivates the distinction between evaluated
and unevaluated contexts. However as a consequence of the preceding
paragraph the previous implementation of get_env meant that when code
relied upon the return type thereof in an unevaluated context the
context became an evaluated context in the body of get_env (and
transitively through all statements therein).

The newly-added unit test highlights this. With the previous
implementation of get_env (which uses automatic return type deduction)
compilation fails. With the new implementation of get_env (which
provides an explicit return type) compilation succeeds.