asioexec::completion_token & ::use_sender: Constrain Initiate
Previously async_result<asioexec::completion_token, ...>:: and
async_result<asioexec::use_sender, ...>::initiate:

- Used automatic return type deduction, and
- Were not constrained on any properties of the arguments thereto

This meant that initiating functions of the form were problematic:

  template <typename AsyncReadStream, typename Allocator,
      BOOST_ASIO_COMPLETION_TOKEN_FOR(void (boost::system::error_code,
        std::size_t)) ReadToken = default_completion_token_t<
          typename AsyncReadStream::executor_type>>
  inline auto async_read(AsyncReadStream& s, basic_streambuf<Allocator>& b,
      ReadToken&& token = default_completion_token_t<
        typename AsyncReadStream::executor_type>(),
      constraint_t<
        !is_completion_condition<ReadToken>::value
      > = 0)
    -> decltype(
      async_initiate<ReadToken,
        void (boost::system::error_code, std::size_t)>(
          declval<detail::initiate_async_read_dynbuf_v1<AsyncReadStream>>(),
          token, basic_streambuf_ref<Allocator>(b), transfer_all()));

Note that:

- The above is an actual initiating function from Asio, and
- The return type is defined in terms of the type returned by
  async_initiate

The latter of the two points above means that the compiler must
determine the type of an expression involving async_initiate. Since that
expression transitively defers to async_result<...>::initiate this meant
that the compiler had to determine the return type thereof. Since (as
mentioned above) that function uses automatic return type deduction that
meant the compiler had to substitute into the body thereof. Since errors
in such a context are not SFINAE-friendly this meant that if the
variadic pack of arguments thereto did not consist only of decay-
copyable types there would be a hard compilation error (rather than the
overload simply being removed from the candidate set).

Both newly-added unit tests failed to compile before this change due to
the above.

Added a constraint to both previously-mentioned initiate functions to
address the above.