let_value, _error, & _stopped: Destroy Child Operation State After Completion
Previously the operation states for let_value, let_error, and
let_stopped directly contained the operation states of one or two child
operations. When first connected this number was one, and after the
completion of that first operation and the connection of the sender
returned by the wrapped invocable the number was two. Notably when the
second operation state is connected the first operation has completed.

P3373 proposes changing the above behavior such that after the tuple of
values sent by the first operation is populated the operation state for
the first operation is destroyed. Thereafter that storage can be reused
for the operation state of the second operation thereby:

- Release resources held by the first operation state sooner, and
- Reducing the size of the operation states of let_value, let_error, and
  let_stopped

This commit implements the above-described change.

Helped-by: Eric Niebler <eniebler@nvidia.com>